Broke free Dist, re-did NN, and begun with SubPathOptimization
the new NN (old one is a.k.a Naive) was slower than Naive... no luck there. I took Dist out from the MST class since we might use it in other places. Begun SubPathOptimization which devides a TSP-solution into subpaths which it tries to optimize... don't yet know if it's a good idea or not